error boundary in user info page and App.js
